Small fix strategy (#275)
* add straetgy for small

* delete pri and post conditions and add first strategy

* try fix pri and post condiotions

* Fix cherry pick

* Add points

* add hardware params and strategy for small

* add ax and fix basket strategy

* Fix basket

* fix move

* add move cherry

* add sterategy for small

* Send point to big robot

* add time out for small robot

* add homologation strategy

* add homologation strategy

* Fix homologation strategy for small robot

* Fix homologation for big robot

* fix big staretgy

* Fix UDP bugs (#278)

* fix interpreter and chmod +x

* remove time from client

* add time measurement to esp32

* fix box_driver

* Fix big staregy and cordinates for green side

* Take side cherry

* small strategy

* Small strategy

* Impruve small strategy

* Fix strategy cilj

* Add green side for small

* Add green side for small

* uradi nesto da radi

* uradi nesto da radi

* uradi nesto da radi-100 put

* uradi nesto da radi-100 put

* fis green starytegy

* first day test 1

* fix

* reuse action homologation

* delete actions

* fix

* fix last

* Add new points

* implement try translate and reaction for cherry picking

* increase boundry of obstacles

* increase boundry of obstacles

* increase boundry of obstacles-fix

* type

* fix

* reduce boundry of obstacles-fix

* vel movex

* move_stuck

* calibrate

* revert

* fix

* fix

* fix staretgy by stac

* fix cordinates

* fix small

* reduce boundry of obstacles-fix

* reduce boundry of obstacles-fix

* fix both robot

* add small strategy for small - Florijan

* add point for basket without stack

* try to ome back in home

* Fix points for both robots and add navigate for small

* Fix staretgy

* fix staretgy for small

* include opstacel detection

* delete detection

* delete navigate for small

* add simple blue strategy

* add simple strategyu for small -final

* add simple green strategy

* try to fix small staraegy fir green side

* fix staretgy

* add points for big robot

* Fix 2-3 round

---------

+++ mep3_behavior/include/mep3_behavior/joint_position_command_action.hpp
@@ -87,7 +87,7 @@ namespace mep3_behavior
           BT::InputPort<double>("timeout"),
           BT::OutputPort<double>("feedback_effort"),
           BT::OutputPort<double>("feedback_position"),
-          BT::OutputPort<uint8_t>("result")};
+          BT::OutputPort<double>("result")};
 
       // Dynamic parameters
       for (std::string table : BT::SharedBlackboard::access()->get<std::vector<std::string>>("predefined_tables"))
@@ -100,7 +100,7 @@ namespace mep3_behavior
 
     BT::NodeStatus onResultReceived(const WrappedResult & wr) override
     {
-      setOutput("result", wr.result->result);
+      setOutput("result", (double)wr.result->result);
       setOutput("feedback_effort", wr.result->last_effort);
       setOutput("feedback_position", wr.result->last_position);
       std::cout << "Last result: " << (double)wr.result->result << "; last effort: " << (double)wr.result->last_effort << "; last position: " <<  (double)wr.result->last_position << std::endl;
